requirements.txt
gunicorn
django-heroku
django-environ